requirements.txt
aioredis==1.2.0
asgiref==3.2.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
cloudinary==1.15.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.5.0
Django==1.11.17
django-braces==1.13.0
django-cors-headers==2.5.2
django-material==1.1.1
django-otp==0.5.2
django-render-block==0.6
django-rest-framework-social-oauth2==1.1.0
django-richtextfield==1.4.0
django-template-email==0.1
django-templated-email==2.3.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
googletrans==2.4.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
material==0.1
mock==2.0.0
msgpack==0.6.1
oauthlib==3.0.1
paypalrestsdk==1.13.1
pbr==5.1.3
Pillow==6.0.0
psycopg2-binary==2.7.7
py-translate==1.0.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyotp==2.2.7
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
Twisted==19.2.1
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
